resmgr: stop importing kubernetes/kubelet internals.
Don't import kubernetes/kubelet for namespace and label
constants or milli-CPU conversion constants and functions.
Instead, define our own constants and functions for those
few things we need.

(cherry picked from commit bf7c90209f13b17f5bb404f5856da7bde426a3d0)

+++ pkg/cri/resource-manager/policy/builtin/podpools/podpools-policy.go
@@ -731,7 +731,7 @@ func (p *podpools) pinCpuMem(c cache.Container, cpus cpuset.CPUSet, mems idset.I
 		c.SetCpusetCpus(cpus.String())
 		if reqCpu, ok := c.GetResourceRequirements().Requests[corev1.ResourceCPU]; ok {
 			mCpu := int(reqCpu.MilliValue())
-			c.SetCPUShares(int64(cache.MilliCPUToShares(mCpu)))
+			c.SetCPUShares(int64(cache.MilliCPUToShares(int64(mCpu))))
 		}
 	}
 	if p.ppoptions.PinMemory {
