+++ pkg/cri/resource-manager/policy/builtin/topology-aware/pools.go
@@ -656,7 +656,7 @@ func (p *policy) applyGrant(grant Grant) {
 		//     processes in the same pool. Also the 'data' process should run fine, since
 		//     it does not need to compete for CPU with any other processes in the system
 		//     as long as that allocation is genuinely system-wide exclusive.
-		container.SetCPUShares(int64(cache.MilliCPUToShares(cpuPortion)))
+		container.SetCPUShares(int64(cache.MilliCPUToShares(int64(cpuPortion))))
 	}
 
 	if mems != "" {
